Add second arg
Added TableToBinary(tab, true). The true will cause TableToBinary to read as sequential table even if it is not.
Passing true will result in a smaller return string as it will not write index keys to file. It will also increase the speed on average to 23x compared to util.TableToJSON.
There is no guarantee the table will be correct however. Any non sequential input in it will cause it to count wrong.
It also assumes every sub table is sequential as well.

This change will not affect BinaryToTable.